+++ src/Controller/Dashboard/MarketPlace/Market/ManufacturerController.php
@@ -147,7 +147,7 @@ class ManufacturerController extends AbstractController
     {
         $market = $this->market($request, $user, $em);
 
-        if ($this->isCsrfTokenValid('delete', $request->get('_token'))) {
+        if ($this->isCsrfTokenValid('delete', $request->get('_token')) && !$manufacturer->getMarketProductManufacturers()->count()) {
             $em->remove($manufacturer);
             $em->flush();
         }
